+++ sdk/resourcemanager/azurestackhci/armazurestackhci/arcsettings_client_example_test.go
@@ -12,7 +12,7 @@ import (
 	"log"
 )
 
-// Generated from example definition: 2026-02-01/ConsentAndInstallDefaultExtensions.json
+// Generated from example definition: 2026-04-01-preview/ConsentAndInstallDefaultExtensions.json
 func ExampleArcSettingsClient_ConsentAndInstallDefaultExtensions() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -74,7 +74,7 @@ func ExampleArcSettingsClient_ConsentAndInstallDefaultExtensions() {
 	// }
 }
 
-// Generated from example definition: 2026-02-01/PutArcSetting.json
+// Generated from example definition: 2026-04-01-preview/PutArcSetting.json
 func ExampleArcSettingsClient_Create() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -136,7 +136,7 @@ func ExampleArcSettingsClient_Create() {
 	// }
 }
 
-// Generated from example definition: 2026-02-01/CreateArcIdentity.json
+// Generated from example definition: 2026-04-01-preview/CreateArcIdentity.json
 func ExampleArcSettingsClient_BeginCreateIdentity() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -170,7 +170,7 @@ func ExampleArcSettingsClient_BeginCreateIdentity() {
 	// }
 }
 
-// Generated from example definition: 2026-02-01/DeleteArcSetting.json
+// Generated from example definition: 2026-04-01-preview/DeleteArcSetting.json
 func ExampleArcSettingsClient_BeginDelete() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -196,7 +196,7 @@ func ExampleArcSettingsClient_BeginDelete() {
 	// }
 }
 
-// Generated from example definition: 2026-02-01/GeneratePassword.json
+// Generated from example definition: 2026-04-01-preview/GeneratePassword.json
 func ExampleArcSettingsClient_GeneratePassword() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -224,7 +224,7 @@ func ExampleArcSettingsClient_GeneratePassword() {
 	// }
 }
 
-// Generated from example definition: 2026-02-01/GetArcSetting.json
+// Generated from example definition: 2026-04-01-preview/GetArcSetting.json
 func ExampleArcSettingsClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -293,7 +293,7 @@ func ExampleArcSettingsClient_Get() {
 	// }
 }
 
-// Generated from example definition: 2026-02-01/InitializeDisableProcess.json
+// Generated from example definition: 2026-04-01-preview/InitializeDisableProcess.json
 func ExampleArcSettingsClient_BeginInitializeDisableProcess() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -319,7 +319,7 @@ func ExampleArcSettingsClient_BeginInitializeDisableProcess() {
 	// }
 }
 
-// Generated from example definition: 2026-02-01/ListArcSettingsByCluster.json
+// Generated from example definition: 2026-04-01-preview/ListArcSettingsByCluster.json
 func ExampleArcSettingsClient_NewListByClusterPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -397,7 +397,7 @@ func ExampleArcSettingsClient_NewListByClusterPager() {
 	}
 }
 
-// Generated from example definition: 2026-02-01/reconcileArcSettings.json
+// Generated from example definition: 2026-04-01-preview/reconcileArcSettings.json
 func ExampleArcSettingsClient_BeginReconcile() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -479,7 +479,7 @@ func ExampleArcSettingsClient_BeginReconcile() {
 	// }
 }
 
-// Generated from example definition: 2026-02-01/PatchArcSetting.json
+// Generated from example definition: 2026-04-01-preview/PatchArcSetting.json
 func ExampleArcSettingsClient_Update() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {