+++ sdk/storage/azqueue/CHANGELOG.md
@@ -1,16 +1,20 @@
 ## Release History
 
-### 2.0.2-beta.1 (Unreleased)
+### 2.1.0-beta.2 (Unreleased)
 
 #### Features Added
-* Added support for service version 2026-04-06.
-* Added cross-tenant support for Principal-Bound User Delegation SAS.
 
 #### Breaking Changes
 
 #### Bugs Fixed
 
 #### Other Changes
+
+### 2.1.0-beta.1 (2026-04-23)
+
+#### Features Added
+* Added support for service version 2026-04-06.
+* Added cross-tenant support for Principal-Bound User Delegation SAS.
 
 ### 2.0.1 (2026-01-12)